+++ rin.Framework/Graphics/Descriptors/BufferWrite.cs
@@ -1,18 +1,18 @@
 ﻿namespace rin.Framework.Graphics.Descriptors;
 
-public class BufferWrite(IDeviceBuffer view, BufferType type, ulong size, ulong offset)
+public class BufferWrite(IDeviceBuffer view, BufferType type, int size,int offset)
 {
     public readonly IDeviceBuffer View = view;
     public readonly BufferType Type = type;
-    public readonly ulong Offset = offset;
-    public readonly ulong Size = size;
+    public readonly int Offset = offset;
+    public readonly int Size = size;
 
     public BufferWrite(IDeviceBuffer view, BufferType type) : this(view,type,view.Size,0)
     {
         
     }
     
-    public BufferWrite(IDeviceBuffer view, BufferType type,ulong size) : this(view,type,size,0)
+    public BufferWrite(IDeviceBuffer view, BufferType type,int size) : this(view,type,size,0)
     {
         
     }
