+++ rin.Framework/Graphics/FrameGraph/IGraphConfig.cs
@@ -5,7 +5,7 @@ namespace rin.Framework.Graphics.FrameGraph;
 public interface IGraphConfig
 {
     public uint CreateImage(uint width, uint height, ImageFormat format, ImageLayout initialLayout = ImageLayout.Undefined);
-    public uint AllocateBuffer(ulong size);
+    public uint AllocateBuffer(int size);
     public uint Read(uint resourceId);
     public uint Write(uint resourceId);
