+++ rin.Framework/Graphics/FrameGraph/MemoryResourceDescriptor.cs
@@ -2,5 +2,5 @@ namespace rin.Framework.Graphics.FrameGraph;
 
 public struct MemoryResourceDescriptor : IResourceDescriptor
 {
-    public required ulong Size;
+    public required int Size;
 }
